requirements.txt
requests
clickhouse-driver
python-dotenv